refactor: batch UPDI instruction sends, u16 tinyAVR addressing
instr2/3/4/5 chain shared out-of-line bodies replacing per-byte call
sites. send_acked fuses store-phase data send and ACK check. TinyProgrammer
takes u16 flash offsets, matching the 16-bit tinyAVR data space.
UPDI stack 1302 -> 1094 B, cellprog 3713 -> 3553 B.

+++ libraries/cellprog/src/session.rs
@@ -193,7 +193,7 @@ impl SessionHandler {
         let Some(data) = self.tx.get(..len) else {
             return self.status_reply(SessionStatus::InvalidAddr, Some(addr));
         };
-        let result = prog.write_flash(u32::from(addr), data);
+        let result = prog.write_flash(addr, data);
         let status = result.map_or_else(|err| status_of(&err), |()| SessionStatus::Ok);
         self.status_reply(status, Some(addr))
     }
@@ -216,7 +216,7 @@ impl SessionHandler {
         // build the reply frame from it into `tx`.
         let Self { rx, tx, .. } = self;
         let data = rx.get_mut(3..3 + len).unwrap_or(&mut []);
-        let status = match prog.read_flash(u32::from(addr), data) {
+        let status = match prog.read_flash(addr, data) {
             Ok(()) => SessionStatus::Ok,
             Err(err) => status_of(&err),
         };
